requirements.txt
# Database
sqlalchemy>=1.4.0,<1.5.0
alembic>=1.7.0,<1.8.0
psycopg2-binary>=2.9.1,<2.10.0

# API Framework
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
pydantic-settings>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.0.6
email-validator>=1.1.3,<1.2.0

# Utilities
implicit>=0.6.2,<0.7.0
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
python-slugify>=8.0.1
python-dotenv>=0.19.0,<0.20.0
numpy>=1.26.0,<1.27.0
pandas>=2.2.0,<2.3.0
scikit-learn>=1.3.0,<1.4.0
scipy>=1.11.0,<1.12.0
joblib>=1.0.1,<1.1.0
transformers>=4.36.0,<4.37.0
torch>=2.1.0,<2.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0,<0.24.0
httpx>=0.24.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
